Add windowing
Added current random permutation of the input to the config dict,
along with a parameter for context window size.

Modified placeholder_inputs() to create a context window in the
noisy inputs, and changed method signature to accept the number of
feature dimensions and context window size, instead of batch size,
which was unused. Also modified all calls to this method as
appropriate.

Moved permutation of the noisy inputs into the batch preparation
part of fill_feed_dict(), since windowing now requires the stacking
up of the context frame sequences.